projects: max11205pmb1: use the example infrastructure
Migrate max11205pmb1 to the shared examples infrastructure (examples.mk):
replace the Makefile BASIC_EXAMPLE/IIO_EXAMPLE toggles with EXAMPLE := iio
+ include examples.mk, split the iio sources into
src/examples/iio/example.mk, drop examples_src.mk and platform_includes.h,
and shorten the example directories (basic_example -> basic,
iio_example -> iio). main.c keeps the NVIC IRQ setup (needed by both
examples) and now defers to example_main(); the basic example owns its
UART stdio init and the iio example includes iio_app.h directly (it used
to come in transitively via platform_includes.h). Update the README to
select the example with make EXAMPLE=<name> on the command line.
