+++ projects/max11205pmb1/Makefile
@@ -1,8 +1,8 @@
-# Select the example you want to enable by choosing y for enabling and n for disabling
-BASIC_EXAMPLE=n
-IIO_EXAMPLE=y
+EXAMPLE ?= iio
 
 include ../../tools/scripts/generic_variables.mk
+
+include ../../tools/scripts/examples.mk
 
 include src.mk
