+++ projects/max11205pmb1/src/common/common_data.h
@@ -33,7 +33,7 @@
 #ifndef __COMMON_DATA_H__
 #define __COMMON_DATA_H__
 
-#include "platform_includes.h"
+#include "parameters.h"
 #include "max11205.h"
 
 extern struct no_os_uart_init_param max11205_uart_ip;
