--- projects/max11205pmb1/src/examples/examples_src.mk
@@ -1,19 +1,0 @@
-ifeq (y,$(strip $(BASIC_EXAMPLE)))
-CFLAGS += -DBASIC_EXAMPLE
-SRCS += $(PROJECT)/src/examples/basic_example/basic_example.c
-INCS += $(PROJECT)/src/examples/basic_example/basic_example.h
-endif
-
-ifeq (y,$(strip $(IIO_EXAMPLE)))
-IIOD=y
-CFLAGS += -DIIO_EXAMPLE
-SRCS += $(PROJECT)/src/examples/iio_example/iio_example.c
-INCS += $(PROJECT)/src/examples/iio_example/iio_example.h
-endif
-
-ifeq (y,$(strip $(IIOD)))
-SRC_DIRS += $(NO-OS)/iio/iio_app
-
-INCS += $(DRIVERS)/adc/max11205/iio_max11205.h
-SRCS += $(DRIVERS)/adc/max11205/iio_max11205.c
-endif
