+++ projects/max11205pmb1/src/examples/iio/example.mk
@@ -1,0 +1,4 @@
+IIOD = y
+
+INCS += $(DRIVERS)/adc/max11205/iio_max11205.h
+SRCS += $(DRIVERS)/adc/max11205/iio_max11205.c
