--- projects/max11205pmb1/src/examples/iio_example/iio_example.c
@@ -1,95 +1,0 @@
-/***************************************************************************//**
- *   @file   iio_example.c
- *   @brief  Implementation of IIO example for max11205pmb1 project.
- *   @author RBolboac (ramona.bolboaca@analog.com)
-********************************************************************************
- * Copyright 2022(c) Analog Devices, Inc.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice,
- *    this list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * 3. Neither the name of Analog Devices, Inc. nor the names of its
- *    contributors may be used to endorse or promote products derived from this
- *    software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY ANALOG DEVICES, INC. “AS IS” AND ANY EXPRESS OR
- * IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
- * MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO
- * EVENT SHALL ANALOG DEVICES, INC. BE LIABLE FOR ANY DIRECT, INDIRECT,
- * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA,
- * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
- * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
- * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
- * EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*******************************************************************************/
-
-#include "iio_example.h"
-#include "iio_max11205.h"
-#include "common_data.h"
-#include "no_os_util.h"
-
-#define DATA_BUFFER_SIZE 400
-
-uint8_t iio_data_buffer[DATA_BUFFER_SIZE * sizeof(int16_t)];
-
-/***************************************************************************//**
- * @brief IIO example main execution.
- *
- * @return ret - Result of the example execution. If working correctly, will
- *               execute continuously function iio_app_run and will not return.
-*******************************************************************************/
-int iio_example_main()
-{
-	int ret;
-	struct iio_app_desc *app;
-	struct max11205_iio_dev *max11205_iio_desc;
-	struct max11205_iio_dev_init_param max11205_iio_ip;
-	struct no_os_irq_ctrl_desc *max11205_gpio_irq_desc;
-	struct iio_data_buffer accel_buff = {
-		.buff = (void *)iio_data_buffer,
-		.size = DATA_BUFFER_SIZE * sizeof(int16_t)
-	};
-	struct iio_app_init_param app_init_param = { 0 };
-
-	/* Initialize GPIO IRQ controller */
-	ret = no_os_irq_ctrl_init(&max11205_gpio_irq_desc, &max11205_gpio_irq_ip);
-	if (ret)
-		return ret;
-
-	/* Initialize device */
-	max11205_ip.irq_ctrl = max11205_gpio_irq_desc;
-
-	max11205_iio_ip.max11205_dev_init = &max11205_ip;
-	max11205_iio_ip.dev_id = MAX11205A;
-
-	ret = max11205_iio_init(&max11205_iio_desc, &max11205_iio_ip);
-	if (ret)
-		return ret;
-
-	struct iio_app_device iio_devices[] = {
-		{
-			.name = "max11205a",
-			.dev = max11205_iio_desc,
-			.dev_descriptor = max11205_iio_desc->iio_dev,
-			.read_buff = &accel_buff,
-		}
-	};
-
-	app_init_param.devices = iio_devices;
-	app_init_param.nb_devices = NO_OS_ARRAY_SIZE(iio_devices);
-	app_init_param.uart_init_params = max11205_uart_ip;
-
-	ret = iio_app_init(&app, app_init_param);
-	if (ret)
-		return ret;
-
-	return iio_app_run(app);
-}
